@@ -1,18 +1,15 @@
 import MEGADomain
 
-public struct AdsSlotConfig: Equatable {
+public struct AdsSlotConfig: Equatable, Sendable {
     let adsSlot: AdsSlotEntity
     let displayAds: Bool
-    let isAdsCookieEnabled: () async -> Bool
     
     public init(
         adsSlot: AdsSlotEntity,
-        displayAds: Bool,
-        isAdsCookieEnabled: @escaping () async -> Bool
+        displayAds: Bool
     ) {
         self.adsSlot = adsSlot
         self.displayAds = displayAds
-        self.isAdsCookieEnabled = isAdsCookieEnabled
     }
     
     public static func == (lhs: AdsSlotConfig, rhs: AdsSlotConfig) -> Bool {
